# services/copilot-agent/app/main.py
"""FastAPI application entry point for the Clinical Co-Pilot agent service."""

from fastapi import Depends, FastAPI
from fastapi.responses import HTMLResponse, JSONResponse

from app.body_size_limit import BodySizeLimitMiddleware
from app.chat import chat_endpoint
from app.config import get_settings
from app.correlation import CorrelationIdMiddleware, configure_logging
from app.dashboard import dashboard_endpoint
from app.documents import source_document_endpoint
from app.feedback import feedback_endpoint
from app.readiness import ReadinessReport, ReadyResponseBody, compute_readiness
from app.review_page import promote_endpoint, review_queue_endpoint

configure_logging()

CHAT_SHELL_HTML = """<!DOCTYPE html>
<html lang="en">
<head>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width, initial-scale=1">
<title>Clinical Co-Pilot</title>
<style>
  * { box-sizing: border-box; }
  html, body { margin: 0; padding: 0; }
  body {
    display: flex;
    flex-direction: column;
    height: 100dvh;
    font-family: system-ui, sans-serif;
    background: #f5f5f5;
    color: #1a1a1a;
  }
  header {
    padding: 0.75rem 1rem;
    background: #0b5a8a;
    color: #fff;
  }
  header h1 {
    margin: 0;
    font-size: 1.1rem;
  }
  #chat-stream {
    flex: 1;
    overflow-y: auto;
    padding: 1rem;
  }
  #chat-form {
    display: flex;
    gap: 0.5rem;
    padding: 0.75rem;
    border-top: 1px solid #ddd;
    background: #fff;
  }
  #chat-input {
    flex: 1;
    min-width: 0;
    padding: 0.6rem;
    font-size: 1rem;
    resize: none;
  }
  #chat-send {
    padding: 0.6rem 1rem;
    font-size: 1rem;
    background: #0b5a8a;
    color: #fff;
    border: none;
  }

  @media (min-width: 768px) {
    body {
      align-items: center;
    }
    header, #chat-stream, #chat-form {
      width: 100%;
      max-width: 720px;
    }
  }
</style>
</head>
<body>
<header>
<h1>Clinical Co-Pilot</h1>
</header>
<main id="chat-stream" data-testid="chat-stream"></main>
<form id="chat-form">
<textarea id="chat-input" data-testid="chat-input" rows="1"></textarea>
<button id="chat-send" data-testid="chat-send" type="submit">Send</button>
</form>
</body>
</html>
"""


def create_app() -> FastAPI:
    """Build and return the FastAPI application."""
    # Issue 214: version-leads-tag from v2.3.0 onward -- this literal must
    # stay in lockstep with pyproject.toml's [project].version (enforced by
    # tests/test_package_version.py) and the pinned openapi/openapi.json
    # (regenerate via `python scripts/generate_openapi_spec.py`).
    app = FastAPI(title="copilot-agent", version="2.3.0")
    app.add_middleware(CorrelationIdMiddleware)
    # Issue #173: registered AFTER CorrelationIdMiddleware above -- Starlette
    # applies add_middleware in REVERSE order of registration (last added =
    # outermost), so this ends up OUTSIDE it. This makes the CONTENT-LENGTH
    # pre-check reject before a correlation id is ever minted/logged --
    # but NOT the streaming-byte-counter check, which runs inside
    # CorrelationIdMiddleware and so DOES carry a correlation id (measured;
    # see app/body_size_limit.py's module docstring for both paths' actual
    # behaviour, not just the ordering rationale). See app/config.py's
    # DEFAULT_MAX_REQUEST_BODY_BYTES for the cap value.
    app.add_middleware(
        BodySizeLimitMiddleware, max_bytes=get_settings().copilot_max_request_body_bytes
    )

    @app.get("/health")
    def health() -> dict[str, str]:
        return {"status": "ok"}

    @app.get(
        "/ready",
        responses={
            200: {"model": ReadyResponseBody},
            503: {"model": ReadyResponseBody},
        },
    )
    async def ready(report: ReadinessReport = Depends(compute_readiness)) -> JSONResponse:
        body = {
            "status": "ready" if report.ready else "not_ready",
            "checks": {
                name: {"ok": result.ok, "detail": result.detail}
                for name, result in report.checks.items()
            },
        }
        return JSONResponse(status_code=200 if report.ready else 503, content=body)

    @app.get("/chat")
    def chat_shell() -> HTMLResponse:
        return HTMLResponse(content=CHAT_SHELL_HTML)

    app.add_api_route("/dashboard", dashboard_endpoint, methods=["GET"])
    app.add_api_route("/review", review_queue_endpoint, methods=["GET"])
    app.add_api_route("/review/promote", promote_endpoint, methods=["GET"])
    app.add_api_route("/chat", chat_endpoint, methods=["POST"])
    app.add_api_route("/feedback", feedback_endpoint, methods=["POST"], status_code=201)
    app.add_api_route("/documents/{source_id}", source_document_endpoint, methods=["GET"])

    return app


app = create_app()