# services/copilot-agent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "description": "``POST /chat`` request body.",
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "patient_id"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "FeedbackRequest": {
        "description": "``POST /feedback`` request body.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "correlation_id": {
            "minLength": 1,
            "title": "Correlation Id",
            "type": "string"
          },
          "thumb": {
            "$ref": "#/components/schemas/FeedbackThumb"
          }
        },
        "required": [
          "correlation_id",
          "thumb"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "FeedbackResponse": {
        "description": "``POST /feedback`` response body: confirms what was recorded and its linkage.",
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "thumb": {
            "$ref": "#/components/schemas/FeedbackThumb"
          }
        },
        "required": [
          "correlation_id",
          "thumb"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
      "FeedbackThumb": {
        "enum": [
          "up",
          "down"
        ],
        "title": "FeedbackThumb",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ReadyCheckBody": {
        "description": "OpenAPI-facing shape of a single ``GET /ready`` check entry.\n\nDeclared separately from ``CheckResult`` (a plain dataclass): this model\nexists only so ``app.main``'s ``responses=`` route metadata can document\nthe real ``/ready`` body shape for BOTH the 200 and 503 cases -- the\nhandler still builds the body by hand (it returns a ``JSONResponse``,\nnot this model, since the status code varies), so this is spec-accuracy\nonly and changes no runtime behavior.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "detail"
        ],
        "title": "ReadyCheckBody",
        "type": "object"
      },
      "ReadyResponseBody": {
        "description": "OpenAPI-facing shape of the full ``GET /ready`` response body.",
        "properties": {
          "checks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReadyCheckBody"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "ReadyResponseBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "copilot-agent",
    "version": "2.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat": {
      "get": {
        "operationId": "chat_shell_chat_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Chat Shell"
      },
      "post": {
        "operationId": "chat_endpoint_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Endpoint"
      }
    },
    "/dashboard": {
      "get": {
        "operationId": "dashboard_endpoint_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Dashboard Endpoint"
      }
    },
    "/documents/{source_id}": {
      "get": {
        "operationId": "source_document_endpoint_documents__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Document Endpoint"
      }
    },
    "/feedback": {
      "post": {
        "operationId": "feedback_endpoint_feedback_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback Endpoint"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponseBody"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponseBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ready"
      }
    },
    "/review": {
      "get": {
        "operationId": "review_queue_endpoint_review_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Review Queue Endpoint"
      }
    },
    "/review/promote": {
      "get": {
        "operationId": "promote_endpoint_review_promote_get",
        "parameters": [
          {
            "in": "query",
            "name": "correlation_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Correlation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Endpoint"
      }
    }
  }
}